split Python and C implementation
